SW_innovation_Backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
future==0.18.3
geographiclib==2.0
geopy==2.4.0
gevent==23.9.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
nominatim==0.1
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tldextract==5.0.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
zope.event==5.0
zope.interface==6.1
